+++ ServerUDP/Server.h
@@ -48,7 +48,7 @@ public:
 
 	void save();
 
-	int processClientsByPair(bool onlyActive, std::function<int(ConnectedClient&)> handler);
+    size_t processClientsByPair(bool onlyActive, std::function<int(ConnectedClient&)> handler);
 
 	// Instances
